requirements.txt
beautifulsoup4==4.11.1
chalice==1.26.3
discord.py==1.7.3
python-dotenv==0.20.0
pytz==2021.3
requests==2.27.1
tinydb==4.7.0
